+++ packages/thumbprint-react/CHANGELOG.md
@@ -6,6 +6,12 @@
 
 -   [Major] Remove `DatePicker`, `Select`, `ModalDefault`, `ModalDefaultHeader`, `ModalDefaultTitle`, `ModalDefaultDescription`, `ModalDefaultContent`, `ModalDefaultContentFullBleed`, `ModalDefaultFooter`, `ModalDefaultAnimatedWrapper`, `Textarea`, `Input`, `InputIcon`, and `InputClearButton` components. These have all been renamed and continue to exist under the new names. (#566)
 
+## 11.1.1 - 2019-12-13
+
+### Fixed
+
+-   [Patch] The `Checkbox` HTML will no longer include the `value` attribute if the value prop is `undefined`. (#589)
+
 ### Changed
 
 -   [Patch] Rename a few directories in our source files. This doesn't affect consumers. (#566)